@@ -1,16 +1,18 @@
+# frozen_string_literal: true
+
 module TokyoApi
   class Client < Vertebrae::API
     def actionkit
       @actionkit ||= TokyoApi::Actionkit.new(client: self)
     end
 
-    alias_method :action_kit, :actionkit
+    alias action_kit actionkit
 
     def bluestatedigital
       @bluestatedigital ||= TokyoApi::Bluestatedigital.new(client: self)
     end
 
-    alias_method :blue_state_digital, :bluestatedigital
+    alias blue_state_digital bluestatedigital
 
     def campact
       @campact ||= TokyoApi::Campact.new(client: self)
@@ -24,7 +26,6 @@ module TokyoApi
       @expire ||= TokyoApi::Expire.new(client: self)
     end
 
-
     def default_options
       opts = {
         user_agent: 'TokyoApi Gem',
@@ -32,32 +33,27 @@ module TokyoApi
         content_type: 'application/json'
       }
 
-      if initialisation_options.has_key?(:timeout)
-        opts[:connection_options] = {request: {timeout: initialisation_options[:timeout]}}
+      if initialisation_options.key?(:timeout)
+        opts[:connection_options] = { request: { timeout: initialisation_options[:timeout] } }
       end
 
       opts
     end
 
     def request(method, path, params, options) # :nodoc:
-      if !::Vertebrae::Request::METHODS.include?(method)
-        raise ArgumentError, "unknown http method: #{method}"
-      end
+      raise ArgumentError, "unknown http method: #{method}" unless ::Vertebrae::Request::METHODS.include?(method)
 
-      unless connection.configuration.prefix.blank?
-        path =  connection.configuration.prefix + '/' + path
-      end
+      path = "#{connection.configuration.prefix}/#{path}" unless connection.configuration.prefix.blank?
 
       ::Vertebrae::Base.logger.debug "EXECUTED: #{method} - #{path} with #{params} and #{options}"
 
       connection.connection.send(method) do |request|
-
         case method.to_sym
-          when *(::Vertebrae::Request::METHODS - ::Vertebrae::Request::METHODS_WITH_BODIES)
-            request.url(path, params)
-          when *::Vertebrae::Request::METHODS_WITH_BODIES
-            request.path = path
-            request.body = extract_data_from_params(params) unless params.empty?
+        when *(::Vertebrae::Request::METHODS - ::Vertebrae::Request::METHODS_WITH_BODIES)
+          request.url(path, params)
+        when *::Vertebrae::Request::METHODS_WITH_BODIES
+          request.path = path
+          request.body = extract_data_from_params(params) unless params.empty?
         end
       end
     end
@@ -67,7 +63,8 @@ module TokyoApi
         builder.use Faraday::Request::Multipart
         builder.use Faraday::Request::UrlEncoded
         if connection.configuration.authenticated?
-          builder.use Faraday::Request::BasicAuthentication, connection.configuration.username, connection.configuration.password
+          builder.use Faraday::Request::BasicAuthentication, connection.configuration.username,
+                      connection.configuration.password
         end
 
         builder.use Faraday::Response::Logger if ENV['DEBUG']
