@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module TokyoApi
   class Identity < Base
     def base_path
@@ -12,21 +14,16 @@ module TokyoApi
       path = "/#{normalized_base_path}user/#{url_escape(id)}"
 
       params = []
-      unless required_fields.nil?
-        params << required_fields_param(required_fields)
-      end
+      params << required_fields_param(required_fields) unless required_fields.nil?
 
       if with_subscription_status
         params << 'with_subscription_status=true'
-        additional_subscription_parameters = path_for_subscription_status_params(opt_in_public_ids, minimum_consent_level, encrypted)
-        unless additional_subscription_parameters.blank?
-          params << additional_subscription_parameters
-        end
+        additional_subscription_parameters = path_for_subscription_status_params(opt_in_public_ids,
+                                                                                 minimum_consent_level, encrypted)
+        params << additional_subscription_parameters unless additional_subscription_parameters.blank?
       end
 
-      if params.any?
-        path << "?#{params.join('&')}"
-      end
+      path << "?#{params.join('&')}" if params.any?
 
       path
     end
@@ -34,7 +31,9 @@ module TokyoApi
     def subscription_status_path(id, opt_in_public_ids: nil, minimum_consent_level: nil, encrypted: nil)
       raise 'must provide opt_in_public_ids' if opt_in_public_ids.nil?
 
-      "/#{normalized_base_path}subscription_status/#{url_escape(id)}?#{path_for_subscription_status_params(opt_in_public_ids, minimum_consent_level, encrypted)}"
+      "/#{normalized_base_path}subscription_status/#{url_escape(id)}?#{path_for_subscription_status_params(
+        opt_in_public_ids, minimum_consent_level, encrypted
+      )}"
     end
 
     private
@@ -42,17 +41,11 @@ module TokyoApi
     def path_for_subscription_status_params(opt_in_public_ids, minimum_consent_level, encrypted = nil)
       path = ''
 
-      if opt_in_public_ids.present?
-        path = "opt_in_public_ids=#{url_escape(opt_in_public_ids.join(','))}"
-      end
+      path = "opt_in_public_ids=#{url_escape(opt_in_public_ids.join(','))}" if opt_in_public_ids.present?
 
-      if minimum_consent_level
-        path << "&minimum_consent_level=#{url_escape(minimum_consent_level)}"
-      end
+      path << "&minimum_consent_level=#{url_escape(minimum_consent_level)}" if minimum_consent_level
 
-      if encrypted
-        path << "&encrypted=1"
-      end
+      path << '&encrypted=1' if encrypted
 
       path
     end
